renamed asm_interface to interface for consistency and removed the current_function field in function context as it wasnt being used
